requirements.txt
python-gitlab==4.11.1
python-dotenv==1.0.1  # Include only if using .env files
requests==2.32.3
httpx==0.27.2